resolviendo inicializar interación
